resistor-colors: rename to resistor-color-duo
RE: https://github.com/exercism/problem-specifications/issues/1495
