+++ config.json
@@ -1556,7 +1556,7 @@
       ]
     },
     {
-      "slug": "resistor-colors",
+      "slug": "resistor-color-duo",
       "uuid": "089f06a6-0759-479c-8c00-d699525a1e22",
       "core": false,
       "unlocked_by": null,
