+++ exercises/resistor-color-duo/example.py
