+++ exercises/resistor-color-duo/resistor_color_duo.py
